+++ examples/sveltekit/src/hooks.server.ts
@@ -1,5 +1,4 @@
 import { auth } from '$lib/server/lucia';
 import { handleHooks } from '@lucia-auth/sveltekit';
-import type { Handle } from '@sveltejs/kit';
 
-export const handle: Handle = handleHooks(auth);
+export const handle = handleHooks(auth);