+++ test-apps/nextjs/pages/login/index.tsx
@@ -1,6 +1,4 @@
-import React, { useState } from "react";
 import Link from "next/link";
-import { useRouter } from "next/router";
 import { auth } from "../../lib/lucia";
 import { AuthRequest } from "@lucia-auth/nextjs";
 
@@ -10,7 +8,7 @@ export const getServerSideProps = async (
 	context: GetServerSidePropsContext
 ): Promise<GetServerSidePropsResult<{}>> => {
 	const authRequest = new AuthRequest(auth, context.req, context.res);
-	const session = await authRequest.getSession();
+	const session = await authRequest.validate();
 	if (session) {
 		return {
 			redirect: {