+++ test-apps/sveltekit/src/routes/signup/+page.server.ts
@@ -39,7 +39,7 @@ import { redirect } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ locals }) => {
-	const session = await locals.getSession();
+	const session = await locals.validate();
 	if (session) throw redirect(302, '/');
 	return {};
 };