+++ packages/astro/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.4.0
+
+- [Breaking] Renamed `authRequest.validate()` to `AuthRequest.validate()`, `authRequest.validateUser()` to `AuthRequest.validateUser()`
+
 ## 0.3.0
 
 - [Breaking] Requires `lucia-auth` 0.3.0