+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.4.0
+
+- [Breaking] Renamed `authRequest.validate()` to `AuthRequest.validate()`, `authRequest.validateUser()` to `AuthRequest.validateUser()`
+
 ## 0.3.0
 
 - [Breaking] Requires `lucia-auth` 0.3.0
@@ -14,7 +18,7 @@
 
 ## 0.1.1
 
-- [Breaking] Removed `AuthRequest().clearSession()`
+- [Breaking] Removed `AuthRequest.clearSession()`
 - [Breaking] `signOut()` will throw an error on failed session invalidation
 - `AuthRequest().setSession()` can be called using `null` to delete all session cookies
 - Update `lucia-auth` to 0.1.4